+++ src/main/java/org/ow2/proactive/catalog/callgraph/CatalogObjectCallGraphPDFGenerator.java
@@ -26,31 +26,21 @@
 package org.ow2.proactive.catalog.callgraph;
 
 import java.io.ByteArrayOutputStream;
-import java.io.File;
-import java.io.IOException;
-import java.util.HashSet;
 import java.util.List;
 import java.util.Optional;
-import java.util.Set;
-import java.util.stream.Collectors;
 
-import org.apache.commons.lang3.StringUtils;
 import org.apache.pdfbox.pdmodel.PDDocument;
 import org.apache.pdfbox.pdmodel.PDPage;
-import org.apache.pdfbox.pdmodel.font.PDType0Font;
 import org.ow2.proactive.catalog.dto.CatalogObjectMetadata;
-import org.ow2.proactive.catalog.dto.Metadata;
 import org.ow2.proactive.catalog.report.HeadersBuilder;
 import org.ow2.proactive.catalog.service.CatalogObjectService;
 import org.ow2.proactive.catalog.service.exception.PDFGenerationException;
+import org.ow2.proactive.catalog.util.ReportGeneratorHelper;
 import org.ow2.proactive.catalog.util.SeparatorUtility;
-import org.ow2.proactive.catalog.util.parser.WorkflowParser;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
 import be.quodlibet.boxable.BaseTable;
-import be.quodlibet.boxable.utils.FontUtils;
 
 
 /**
@@ -65,18 +55,6 @@ public class CatalogObjectCallGraphPDFGenerator {
 
     private static final String MAIN_TITLE = "ProActive Call Graph Report";
 
-    @Value("${pa.catalog.pdf.report.ttf.font.path}")
-    private String ttfFontPath;
-
-    @Value("${pa.catalog.pdf.report.ttf.font.bold.path}")
-    private String ttfFontBoldPath;
-
-    @Value("${pa.catalog.pdf.report.ttf.font.italic.path}")
-    private String ttfFontItalicPath;
-
-    @Value("${pa.catalog.pdf.report.ttf.font.bold.italic.path}")
-    private String ttfFontBoldItalicPath;
-
     @Autowired
     private HeadersBuilder headersBuilder;
 
@@ -84,7 +62,10 @@ public class CatalogObjectCallGraphPDFGenerator {
     private SeparatorUtility separatorUtility;
 
     @Autowired
-    private TableCallGraphsBuilder tableGraphPathBuilder;
+    private TableCallGraphsBuilder tableCallGraphsBuilder;
+
+    @Autowired
+    private ReportGeneratorHelper reportGeneratorHelper;
 
     public byte[] generatePdfImage(List<CatalogObjectMetadata> catalogObjectMetadataList, Optional<String> kind,
             Optional<String> contentType, CatalogObjectService catalogObjectService) {
@@ -93,16 +74,17 @@ public class CatalogObjectCallGraphPDFGenerator {
                 PDDocument doc = new PDDocument()) {
 
             // Build call graph
-            CallGraphHolder globalCallGraph = buildCatalogCallGraph(catalogObjectMetadataList, catalogObjectService);
+            CallGraphHolder globalCallGraph = reportGeneratorHelper.buildCatalogCallGraph(catalogObjectMetadataList,
+                                                                                          catalogObjectService);
 
             //Load font for all languages
-            setFontToUse(doc);
+            reportGeneratorHelper.setFontToUse(doc);
 
             // Initialize Document
-            PDPage page = addNewPage(doc);
+            PDPage page = reportGeneratorHelper.addNewPage(doc);
 
             // Initialize table
-            BaseTable table = initializeTable(doc, MARGIN, page);
+            BaseTable table = reportGeneratorHelper.initializeTable(doc, MARGIN, page);
 
             // Create Header row
             headersBuilder.createMainHeader(table, MAIN_TITLE);
@@ -110,13 +92,13 @@ public class CatalogObjectCallGraphPDFGenerator {
             // Create Header row
 
             headersBuilder.createInfoHeader(table,
-                                            extractBucketSet(globalCallGraph),
-                                            extractObjectSet(globalCallGraph),
+                                            reportGeneratorHelper.extractBucketSet(globalCallGraph),
+                                            reportGeneratorHelper.extractObjectSet(globalCallGraph),
                                             kind,
                                             contentType);
 
-            //Create graph path table
-            tableGraphPathBuilder.buildCallGraphsTable(globalCallGraph, table);
+            //Create call graphs table
+            tableCallGraphsBuilder.buildCallGraphsTable(globalCallGraph, table);
 
             table.draw();
 
@@ -127,104 +109,5 @@ public class CatalogObjectCallGraphPDFGenerator {
         } catch (Exception e) {
             throw new PDFGenerationException(e);
         }
-    }
-
-    private List<String> collectDependsOnCatalogObjects(CatalogObjectMetadata catalogObjectMetadata) {
-        return catalogObjectMetadata.getMetadataList()
-                                    .stream()
-                                    .filter(metadata -> metadata.getLabel()
-                                                                .equals(WorkflowParser.ATTRIBUTE_DEPENDS_ON_LABEL))
-                                    .map(Metadata::getKey)
-                                    .collect(Collectors.toList());
-
-    }
-
-    private CallGraphHolder buildCatalogCallGraph(List<CatalogObjectMetadata> catalogObjectMetadataList,
-            CatalogObjectService catalogObjectService) {
-
-        CallGraphHolder callGraphHolder = new CallGraphHolder();
-        for (CatalogObjectMetadata catalogObjectMetadata : catalogObjectMetadataList) {
-            List<String> dependsOnCatalogObjects = collectDependsOnCatalogObjects(catalogObjectMetadata);
-            if (!dependsOnCatalogObjects.isEmpty()) {
-                GraphNode callingCatalogObject = callGraphHolder.addNode(catalogObjectMetadata.getBucketName(),
-                                                                         catalogObjectMetadata.getName(),
-                                                                         catalogObjectMetadata.getKind(),
-                                                                         true);
-                String bucketName;
-                String objectName;
-                boolean isCatalogObjectExist;
-                String objectKind;
-                GraphNode calledCatalogObject;
-                for (String dependsOnCatalogObject : dependsOnCatalogObjects) {
-                    bucketName = separatorUtility.getSplitBySeparator(dependsOnCatalogObject).get(0);
-                    objectName = separatorUtility.getSplitBySeparator(dependsOnCatalogObject).get(1);
-                    isCatalogObjectExist = catalogObjectService.isDependsOnObjectExistInCatalog(bucketName,
-                                                                                                objectName,
-                                                                                                WorkflowParser.LATEST_VERSION);
-                    objectKind = isCatalogObjectExist ? catalogObjectService.getCatalogObjectMetadata(bucketName,
-                                                                                                      objectName)
-                                                                            .getKind()
-                                                      : "N/A";
-                    calledCatalogObject = callGraphHolder.addNode(bucketName,
-                                                                  objectName,
-                                                                  objectKind,
-                                                                  isCatalogObjectExist);
-                    callGraphHolder.addDependsOnEdge(callingCatalogObject, calledCatalogObject);
-                }
-            }
-        }
-        return callGraphHolder;
-    }
-
-    private void setFontToUse(PDDocument doc) throws IOException {
-        FontUtils.setSansFontsAsDefault(doc);
-        addFontTypeIfFileExists(doc, ttfFontPath, "font");
-        addFontTypeIfFileExists(doc, ttfFontBoldPath, "fontBold");
-        addFontTypeIfFileExists(doc, ttfFontItalicPath, "fontItalic");
-        addFontTypeIfFileExists(doc, ttfFontBoldItalicPath, "fontBoldItalic");
-    }
-
-    private void addFontTypeIfFileExists(PDDocument doc, String path, String fontType) throws IOException {
-        if (!StringUtils.isEmpty(path) && new File(path).exists()) {
-            FontUtils.getDefaultfonts().put(fontType, PDType0Font.load(doc, new File(path)));
-        }
-    }
-
-    private BaseTable initializeTable(PDDocument doc, float margin, PDPage page) throws IOException {
-        float tableWidth = page.getMediaBox().getWidth() - (2 * margin);
-        float yStartNewPage = page.getMediaBox().getHeight() - (2 * margin);
-        boolean drawContent = true;
-        boolean drawLines = true;
-        float yStart = yStartNewPage;
-        float bottomMargin = 70;
-        return new BaseTable(yStart,
-                             yStartNewPage,
-                             bottomMargin,
-                             tableWidth,
-                             margin,
-                             doc,
-                             page,
-                             drawLines,
-                             drawContent);
-    }
-
-    private PDPage addNewPage(PDDocument doc) {
-        PDPage page = new PDPage();
-        doc.addPage(page);
-        return page;
-    }
-
-    private Set<String> extractBucketSet(CallGraphHolder callGraphHolder) {
-        Set<String> bucketSet = new HashSet<>();
-        callGraphHolder.nodeSet().forEach(graphNode -> bucketSet.add(graphNode.getBucketName()));
-        return bucketSet;
-
-    }
-
-    private Set<String> extractObjectSet(CallGraphHolder callGraphHolder) {
-        Set<String> objectSet = new HashSet<>();
-        callGraphHolder.nodeSet().forEach(graphNode -> objectSet.add(graphNode.getObjectName()));
-        return objectSet;
-
     }
 }
