+++ src/main/java/org/ow2/proactive/catalog/report/CatalogObjectReportPDFGenerator.java
@@ -26,46 +26,26 @@
 package org.ow2.proactive.catalog.report;
 
 import java.io.ByteArrayOutputStream;
-import java.io.File;
-import java.io.IOException;
 import java.util.Optional;
 import java.util.Set;
 
-import org.apache.commons.lang3.StringUtils;
 import org.apache.pdfbox.pdmodel.PDDocument;
 import org.apache.pdfbox.pdmodel.PDPage;
-import org.apache.pdfbox.pdmodel.font.PDType0Font;
 import org.ow2.proactive.catalog.dto.CatalogObjectMetadata;
 import org.ow2.proactive.catalog.service.exception.PDFGenerationException;
+import org.ow2.proactive.catalog.util.ReportGeneratorHelper;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
 import be.quodlibet.boxable.BaseTable;
-import be.quodlibet.boxable.utils.FontUtils;
 
 
 @Component
 public class CatalogObjectReportPDFGenerator {
 
-    private final static float MARGIN = 10;
+    private static final float MARGIN = 10;
 
     private static final String MAIN_TITLE = "ProActive Catalog Report";
-
-    @Value("${pa.scheduler.url}")
-    private String schedulerUrl;
-
-    @Value("${pa.catalog.pdf.report.ttf.font.path}")
-    private String ttfFontPath;
-
-    @Value("${pa.catalog.pdf.report.ttf.font.bold.path}")
-    private String ttfFontBoldPath;
-
-    @Value("${pa.catalog.pdf.report.ttf.font.italic.path}")
-    private String ttfFontItalicPath;
-
-    @Value("${pa.catalog.pdf.report.ttf.font.bold.italic.path}")
-    private String ttfFontBoldItalicPath;
 
     @Autowired
     private TableDataBuilder tableDataBuilder;
@@ -73,20 +53,23 @@ public class CatalogObjectReportPDFGenerator {
     @Autowired
     private HeadersBuilder headersBuilder;
 
+    @Autowired
+    private ReportGeneratorHelper reportGeneratorHelper;
+
     public byte[] generatePDF(Set<CatalogObjectMetadata> orderedObjectsPerBucket, Optional<String> kind,
             Optional<String> contentType) {
 
         try (ByteArrayOutputStream byteArrayOutputStream = new ByteArrayOutputStream();
-                PDDocument doc = new PDDocument();) {
+                PDDocument doc = new PDDocument()) {
 
             //Load font for all languages
-            setFontToUse(doc);
+            reportGeneratorHelper.setFontToUse(doc);
 
             // Initialize Document
-            PDPage page = addNewPage(doc);
+            PDPage page = reportGeneratorHelper.addNewPage(doc);
 
             // Initialize table
-            BaseTable table = initializeTable(doc, MARGIN, page);
+            BaseTable table = reportGeneratorHelper.initializeTable(doc, MARGIN, page);
 
             // Create Header row
             headersBuilder.createMainHeader(table, MAIN_TITLE);
@@ -106,44 +89,6 @@ public class CatalogObjectReportPDFGenerator {
         } catch (Exception e) {
             throw new PDFGenerationException(e);
         }
-    }
-
-    private void setFontToUse(PDDocument doc) throws IOException {
-        FontUtils.setSansFontsAsDefault(doc);
-        addFontTypeIfFileExists(doc, ttfFontPath, "font");
-        addFontTypeIfFileExists(doc, ttfFontBoldPath, "fontBold");
-        addFontTypeIfFileExists(doc, ttfFontItalicPath, "fontItalic");
-        addFontTypeIfFileExists(doc, ttfFontBoldItalicPath, "fontBoldItalic");
-    }
-
-    private void addFontTypeIfFileExists(PDDocument doc, String path, String fontType) throws IOException {
-        if (!StringUtils.isEmpty(path) && new File(path).exists()) {
-            FontUtils.getDefaultfonts().put(fontType, PDType0Font.load(doc, new File(path)));
-        }
-    }
-
-    private BaseTable initializeTable(PDDocument doc, float margin, PDPage page) throws IOException {
-        float tableWidth = page.getMediaBox().getWidth() - (2 * margin);
-        float yStartNewPage = page.getMediaBox().getHeight() - (2 * margin);
-        boolean drawContent = true;
-        boolean drawLines = true;
-        float yStart = yStartNewPage;
-        float bottomMargin = 70;
-        return new BaseTable(yStart,
-                             yStartNewPage,
-                             bottomMargin,
-                             tableWidth,
-                             margin,
-                             doc,
-                             page,
-                             drawLines,
-                             drawContent);
-    }
-
-    private PDPage addNewPage(PDDocument doc) {
-        PDPage page = new PDPage();
-        doc.addPage(page);
-        return page;
     }
 
 }
